Add method segmentInside(Rect) to Line and slightly refactor.
This method returns the portion of the Line that is inside Rect,
or an empty boost::optional if the Line doesn't intersect the Rect.

This was previously found in conicsec.cpp and path-cairo.cpp.
However, the implementations there were hard to understand. This new one
is very simple, and based on the new method root(). Methods for
dealing with the equation representation of Line were also improved.