Improve the robustness of boolops.
When constructing the intersection graph, ensure that when
both sides of an intersection have the same winding parity,
this is also the case for the neighbor.

Boolops will now give correct results for shapes with overlapping
edges, but the shapes may contain unnecessary extra nodes.
Avoiding those will require more substantial modifications and
explicit detection of overlapping fragments.