Embed googletest in source tree.
The make dist rules are not set up to handle submodules, so they omit
googletest and the subsequent build breaks.  Since googletest is our
only submodule, and since it hardly ever changes, simplest solution is
to just include its source code directly in the tree.

A better solution would be to use googletest as an external dependency,
however it appears the packaged version in Ubuntu is not an easily drop
in replacement for what is currently in use.  Inkscape uses a script to
download the googletest code as a workaround, however it is also not a
drop in replacement here and would take some adjustments to make work.